Renamed sessionList to sessions "session" Redux stuff is always in a "session" file
